# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de renta de herramientas con aceptación mutua (apretón de manos digital).\n\n## Flujo de renta\n1. Propietario registra herramienta con precio\n2. Solicitante solicita renta → estado **pending**\n3. Ambos llaman `/confirm-delivery` → estado **active**\n4. Ambos llaman `/confirm-return` → estado **completed** (fondos liberados)",
        "title": "Tool Rental API",
        "contact": {
            "name": "Angel Chame",
            "email": "angelchame6@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/owner/tools": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista las herramientas del propietario autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "herramientas"
                ],
                "summary": "Mis herramientas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ToolResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/pricing": {
            "get": {
                "description": "Calcula el precio diario mínimo sugerido basado en el valor estimado de la herramienta (50% del valor en 30 días)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "herramientas"
                ],
                "summary": "Precio sugerido",
                "parameters": [
                    {
                        "type": "number",
                        "example": 1000,
                        "description": "Valor estimado de la herramienta",
                        "name": "estimated_value",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/input.PricingSuggestion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/tools": {
            "get": {
                "description": "Lista todas las herramientas. Acepta filtros por disponibilidad, categoría y búsqueda por nombre",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "herramientas"
                ],
                "summary": "Catálogo de herramientas",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Solo herramientas disponibles",
                        "name": "available",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por categoría",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Buscar por nombre o descripción",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ToolResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea una nueva herramienta. El precio diario mínimo se calcula automáticamente (50% del valor estimado en 30 días)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "herramientas"
                ],
                "summary": "Registrar herramienta",
                "parameters": [
                    {
                        "description": "Datos de la herramienta",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateToolRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ToolResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/tools/{id}": {
            "get": {
                "description": "Devuelve los datos de una herramienta incluyendo precio sugerido",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "herramientas"
                ],
                "summary": "Detalle de herramienta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la herramienta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ToolResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Modifica los campos de una herramienta. Solo el propietario dueño puede hacerlo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "herramientas"
                ],
                "summary": "Actualizar herramienta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la herramienta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateToolRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ToolResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina una herramienta. Solo el propietario dueño puede hacerlo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "herramientas"
                ],
                "summary": "Eliminar herramienta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la herramienta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MsgResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/tools/{id}/photo": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sube una foto desde la cámara o galería. Enviar como multipart/form-data con el campo \"photo\"",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "herramientas"
                ],
                "summary": "Subir foto de herramienta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la herramienta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Archivo de imagen",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ToolResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateToolRequest": {
            "type": "object",
            "required": [
                "daily_rate",
                "estimated_value",
                "name"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "example": "Construcción"
                },
                "daily_rate": {
                    "type": "number",
                    "example": 50
                },
                "description": {
                    "type": "string",
                    "example": "Taladro percutor ideal para concreto"
                },
                "estimated_value": {
                    "type": "number",
                    "example": 2500
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2,
                    "example": "Taladro Bosch 800W"
                }
            }
        },
        "dto.ErrResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "mensaje de error"
                }
            }
        },
...
        "dto.MsgResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operación exitosa"
                }
            }
        },
...
        "dto.ToolResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "daily_rate": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "estimated_value": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "is_available": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "suggested_min_daily_rate": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateToolRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "daily_rate": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "estimated_value": {
                    "type": "number"
                },
                "is_available": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                }
            }
        },
...
        "input.PricingSuggestion": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "estimated_value": {
                    "type": "number"
                },
                "minimum_daily_rate": {
                    "type": "number"
                },
                "suggested_daily_rate": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Pegar el token así: **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}